[nrf fromtree] Bluetooth: Audio: Add packing to unicast group create
Add the ISO packing field when creating a unicast group.

This refactors the structure of the unicast group create,
as it now takes both the packing as a group parameter, as well
as an array of stream-specific parameters.

Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
(cherry picked from commit 4e3205d2386bf92884d8dca1bf66d4fa7267783a)
Signed-off-by: Jui-Chou Chung <jui-chou.chung@nordicsemi.no>